Factor out a FunctionsList component for rendering function docs
